app/requirements.txt
flask 
flask-smorest
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
bcrypt
osmnx
geopandas
folium
matplotlib
mapclassify
geopy
geocoder